Move `TimelineRecords` and `GraphRecords`
This moves the classes to `app/lib` which is where we store all other
autoloaded miscellaneous classes (it also fixes an issue where we've
ended up with duplicate `GraphRecords` classes).
