+++ app/lib/timeline_records.rb
